move MainWindow files to discodos/qt folder
