+++ gui.py
@@ -1,6 +1,6 @@
 import sys
 
-from MainWindow import Ui_MainWindow
+from qt.MainWindow import Ui_MainWindow
 from PyQt5 import QtCore, QtGui, QtWidgets
 from PyQt5.QtCore import Qt
 from discodos.models import Mix, log, Collection
